Split mappers into designated files
